# api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Aegis Trader
  description: AI-assisted, risk-gated automated trading platform. Paper trading only
    in V1; the LLM can only create trade proposals.
  version: 0.1.0
paths:
...
  /api/v1/trades:
    get:
      tags:
      - trades
      summary: List completed trades
      operationId: list_trades_api_v1_trades_get
      security:
      - HTTPBearer: []
      parameters:
      - name: symbol
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Symbol
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradePageSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/trades/{trade_id}:
    get:
      tags:
      - trades
      summary: Trade detail
      operationId: get_trade_api_v1_trades__trade_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: trade_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Trade Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TradePageSchema:
      properties:
        items:
          items:
            $ref: '#/components/schemas/TradeSchema'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
      type: object
      required:
      - items
      - total
      - page
      - page_size
      title: TradePageSchema
    TradeSchema:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        symbol:
          type: string
          title: Symbol
        side:
          $ref: '#/components/schemas/TradeSide'
        quantity:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Quantity
        entry_price:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Entry Price
        exit_price:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Exit Price
        pnl:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Pnl
        fees:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Fees
        return_pct:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Return Pct
        strategy_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Strategy Id
        proposal_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Proposal Id
        order_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Order Id
        opened_at:
          type: string
          format: date-time
          title: Opened At
        closed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Closed At
      type: object
      required:
      - id
      - symbol
      - side
      - quantity
      - entry_price
      - exit_price
      - pnl
      - fees
      - return_pct
      - strategy_id
      - proposal_id
      - order_id
      - opened_at
      - closed_at
      title: TradeSchema
    TradeSide:
      type: string
      enum:
      - BUY
      - SELL
      title: TradeSide
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer